requirements.txt
flask==2.3.3
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
textblob==0.17.1
gensim==4.3.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
plotly==5.15.0
jupyter==1.0.0
ipykernel==6.25.0
google-generativeai==0.3.2
python-dotenv==1.0.0
joblib==1.3.2
tqdm==4.66.1
transformers==4.30.2
torch==2.0.1
sentencepiece==0.1.99
protobuf==3.20.3
accelerate==0.20.3